+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 - better change handling in ientry improves matrixview behaviour
+- better subrow finding fixes replace from file in nested columns
 
 ## 9.0.4 2025/03/20